pyproject.toml
[tool.poetry]
name = "nhp_dwiproc"
version = "0.1.0"
description = "BIDS app for performing diffusion processing of NHP dataset(s)"
authors = ["Jason Kai <jason.kai@childmind.org>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "nhp_dwiproc", from = "src"}]

[tool.poetry.dependencies]
python = "~3.11"
bids2table = "^0.1.0"
pyyaml = "^6.0.1"
niwrap = {git = "https://github.com/childmindresearch/niwrap", subdirectory = "python"}
bidsapp-helper = {git = "https://github.com/kaitj/bidsapp-helper"}
styxdocker = {git = "https://github.com/childmindresearch/styxdocker"}
styxsingularity = {git = "https://github.com/childmindresearch/styxsingularity"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
mypy = "^1.10.1"
pre-commit = "^3.7.1"
pytest-cov = "^5.0.0"
ruff = "^0.5.1"
ipykernel = "^6.29.5"

[tool.poetry.group.docs.dependencies]

[tool.poetry.scripts]
nhp_dwiproc = "nhp_dwiproc.run:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv"
]
line-length = 88
indent-width = 4
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = ["ANN", "D", "E", "F", "I"]
ignore = [
  "ANN101",  # self should not be annotated.
  "ANN102",  # cls should not be annotated.
  "ANN401"
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
